move scatter-lowering and hlo-custom-call-lowering from catalyst dialect to stablehlo folder
