+++ mlir/include/Catalyst/Transforms/Passes.h
@@ -29,13 +29,11 @@ std::unique_ptr<mlir::Pass> createCatalystConversionPass();
 std::unique_ptr<mlir::Pass> createDetensorizeSCFPass();
 std::unique_ptr<mlir::Pass> createDisableAssertionPass();
 std::unique_ptr<mlir::Pass> createGEPInboundsPass();
-std::unique_ptr<mlir::Pass> createHloCustomCallLoweringPass();
 std::unique_ptr<mlir::Pass> createInlineNestedModulePass();
 std::unique_ptr<mlir::Pass> createMemrefCopyToLinalgCopyPass();
 std::unique_ptr<mlir::Pass> createMemrefToLLVMWithTBAAPass();
 std::unique_ptr<mlir::Pass> createQnodeToAsyncLoweringPass();
 std::unique_ptr<mlir::Pass> createRegisterInactiveCallbackPass();
-std::unique_ptr<mlir::Pass> createScatterLoweringPass();
 std::unique_ptr<mlir::Pass> createSplitMultipleTapesPass();
 
 void registerAllCatalystPasses();
