+++ mlir/lib/stablehlo/HloCustomCallPatterns.cpp
@@ -17,9 +17,8 @@
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/IR/PatternMatch.h"
-#include "llvm/Support/Debug.h"
-
 #include "stablehlo/dialect/StablehloOps.h"
+#include "llvm/Support/Debug.h"
 
 #include "Catalyst/IR/CatalystOps.h"
